mod:  refactor user resolver

+++ graph/resolvers/user_resolver.go
@@ -29,12 +29,14 @@ func (r *UserResolver) CreateUser(ctx context.Context, input model.CreateUserInp
 	}
 
 	user, err := r.userService.GetUserByEmail(*input.Email)
+
 	if err == nil && user != nil {
 		return nil, fmt.Errorf("email already exists")
 	}
 	if err != nil && err.Error() != "mongo: no documents in result" {
 		return nil, fmt.Errorf("server: get user by email, details: %w", err)
 	}
+
 	if err != nil && err.Error() == "mongo: no documents in result" {
 		defaultName := "New User"
 		if input.Name == nil || *input.Name == "" {
@@ -50,7 +52,7 @@ func (r *UserResolver) CreateUser(ctx context.Context, input model.CreateUserInp
 			Password: hashedPassword,
 		})
 		if err != nil {
-			return nil, fmt.Errorf("server: create user details: %w", err)
+			return nil, fmt.Errorf("server: create user, details: %w", err)
 		}
 	}
 
@@ -58,25 +60,34 @@ func (r *UserResolver) CreateUser(ctx context.Context, input model.CreateUserInp
 }
 
 func (r *UserResolver) DeleteUser(ctx context.Context, email string) (*models.User, error) {
+	if email == "" {
+		return nil, fmt.Errorf("server: email is required")
+	}
+
 	user, err := r.userService.DeleteUserByEmail(email)
+
+	if err.Error() == "mongo: no documents in result" {
+		return nil, fmt.Errorf("server: no user with specified email")
+	}
+
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("server: delete user by email, details: %w", err)
 	}
 
 	return user, nil
 }
 
-func (r *UserResolver) ID(ctx context.Context, obj *models.User) (string, error) {
-	return obj.Id.Hex(), nil
-}
-
 func (r *UserResolver) Users(ctx context.Context) ([]*models.User, error) {
 	users, err := r.userService.GetUsers()
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("server: get users, details: %w", err)
 	}
 
 	return users, nil
+}
+
+func (r *UserResolver) ID(ctx context.Context, obj *models.User) (string, error) {
+	return obj.Id.Hex(), nil
 }
 
 func (r *UserResolver) User(ctx context.Context, email string) (*models.User, error) {
@@ -85,9 +96,6 @@ func (r *UserResolver) User(ctx context.Context, email string) (*models.User, er
 	}
 
 	user, err := r.userService.GetUserByEmail(email)
-	if user == nil {
-		return nil, fmt.Errorf("server: email already exists")
-	}
 	if err != nil {
 		return nil, err
 	}
